1. Added a more detailed food search 2. Updated a more detailed recipe service 3. Updated some data objects

+++ backend/src/main/java/com/miniproj/miniproj/vo/RecipeVO.java
@@ -8,6 +8,6 @@ import java.util.List;
 public class RecipeVO {
     Integer recipeId;
     String recipeDescription;
-    Integer totalCalories;
+    Double totalCalories;
     List<RecipeFoodVO> foods;
 }
